Simplify all 'Remarks:' elements for constructors
When the expression inside `noexcept` or `explicit` is a single
expression just put it directly in the function declaration, instead of
using "see below". When we retain the "see below" and a separate
Remarks, ensure that the expression is a valid C++ logical expression,
not prose of the form "x is true, and y is true".

Remove "expression inside `noexcept`" for expected(unexpected<G>&&),
because it doesn't have a noexcept-specifier (in neither the class
synopsis nor the detailed description).

Remove all bogus "this constructor is trivial" sentences, and all
normatively useless "is constexpr" sentences.