Change type of unex union member to E, instead of unexpected<E>
This results in changes throughout the wording, mostly simplifications
as we can initialize `unex` directly instead of creating temproaries of
type `unexpected<E>`.

The assignments and emplace functions have been completely rewritten,
using "Equivalent to [code]" for the specifications. This makes many of
the Postconditions, Throws, and Remarks elements unnecessary.

The emplace functions now always construct a new value (destroying any
that was already present) and are only callable if the value is
nothrow-constructible. For cases where the construction might throw you
can just use assignment instead. This simplifies emplace considerably
and allows it to be unconditionally noexcept.

As requested by LWG, it is now ill-formed to instantiate
expewcted<unexpected<T>, E> for any T (where previously it was only
ill-formed if T and E are the same). This results in some changes to the
constraints of expected::expected(U&&) and expected::operator=(U&&). It
also means that the converting constructors and assignment operators of
unexpected could be removed, but that isn't done by this patch. The
implementation never needs to construct an unexpected<E> at all now, so
there are no conversions from unexpected<G> to unexpected<E>.