Changes to std::expected requested by LWG (#83)
* Update date of draft

* Fix headings for bad_expected_access subclauses

* Simplify all 'Remarks:' elements for constructors

When the expression inside `noexcept` or `explicit` is a single
expression just put it directly in the function declaration, instead of
using "see below". When we retain the "see below" and a separate
Remarks, ensure that the expression is a valid C++ logical expression,
not prose of the form "x is true, and y is true".

Remove "expression inside `noexcept`" for expected(unexpected<G>&&),
because it doesn't have a noexcept-specifier (in neither the class
synopsis nor the detailed description).

Remove all bogus "this constructor is trivial" sentences, and all
normatively useless "is constexpr" sentences.

* Model unexpect_t on in_place_t, not nullopt_t

There is no need for the special handling to avoid ambiguities with {}
because std::unexpect is not used for assignment. It's a tag passed to
the constructor only.

* Simplify all cases of direct-list-initialization

Replace "initializes foo as if direct-list-initializing an object of
type T with the expression bar" with simply "direct-list-initializes foo
with bar".

* Remove qualification from expected<cv void> member descriptions

For expected::value() and expected::emplace().

* Combine operator bool() and has_value() for void specialization

Also add missing 'const'. These changes were already done for the
non-void specialization.

* Add missing noexcept to expected synopsis

It was recently added to the detailed descriptions, but missed the synopsis.

* Use `this->has_value()` when `rhs.has_value()` is also mentioned

If there are two objects, use `this` to disambiguate them consistently.

* Remove duplicated deduction guide in the wrong class synopsis

* Replace "selected constructor of X" with "initialization of x"

* Use "the exception specification" instead of "expression inside noexcept"

* Remove normatively meaningless constexpr remarks for equality ops

* Change type of unex union member to E, instead of unexpected<E>

This results in changes throughout the wording, mostly simplifications
as we can initialize `unex` directly instead of creating temproaries of
type `unexpected<E>`.

The assignments and emplace functions have been completely rewritten,
using "Equivalent to [code]" for the specifications. This makes many of
the Postconditions, Throws, and Remarks elements unnecessary.

The emplace functions now always construct a new value (destroying any
that was already present) and are only callable if the value is
nothrow-constructible. For cases where the construction might throw you
can just use assignment instead. This simplifies emplace considerably
and allows it to be unconditionally noexcept.

As requested by LWG, it is now ill-formed to instantiate
expewcted<unexpected<T>, E> for any T (where previously it was only
ill-formed if T and E are the same). This results in some changes to the
constraints of expected::expected(U&&) and expected::operator=(U&&). It
also means that the converting constructors and assignment operators of
unexpected could be removed, but that isn't done by this patch. The
implementation never needs to construct an unexpected<E> at all now, so
there are no conversions from unexpected<G> to unexpected<E>.

* Increment draft revision

* Remove converting constructors and assignments from unexpected

* Remove stray "and" on last item of list

* Move assignment operator should be constrained not deleted

* Add missing template-head to converting assignments

* Remove orphaned reference to table

* Combine lvalue and rvalue overloads of converting ctors and assignments

Introduce UF and GF names to use where the requirements differ. Removes
a lot of duplication.

* Remove duplicate is_void_v<U> constraint

* Add description for operator== for void specialization and unexepcted<E2>

* Fix static preconditions for value_or

Both conditions need to be true, and use Mandates not Remarks.

* Use a single subclause for all operator== overloads